Add committee members is now a separate modal. Committee heads can add members from the committee overview page
